add test cases for git

+++ pkgs/git/control.go
@@ -18,45 +18,57 @@ func CloneRepo(title, repoUrl, repoRef, repoDir string) error {
 	// Try to hack github repo url with proxy url.
 	repoUrl, err := proxy.HackRepoUrl(repoUrl)
 	if err != nil {
-		return err
+		return fmt.Errorf("hack repo url failed: %w", err)
 	}
 
-	switch {
-	case repoRef == "":
-		// Clone default branch.
+	// ============ Clone default branch ============
+	if repoRef == "" {
 		command := fmt.Sprintf("git clone %s --recursive %s", repoUrl, repoDir)
-		if err := cmd.NewExecutor(title, command).Execute(); err != nil {
-			return err
-		}
+		return cmd.NewExecutor(title, command).Execute()
+	}
 
-	case CheckIfRemoteBranch(repoUrl, repoRef), CheckIfRemoteTag(repoUrl, repoRef):
-		// Clone specific branch or tag.
-		command := fmt.Sprintf("git clone --branch %s %s --depth 1 --recursive %s", repoRef, repoUrl, repoDir)
-		if err := cmd.NewExecutor(title, command).Execute(); err != nil {
-			return err
-		}
+	// ============ Clone specific branch ============
+	isBranch, err := CheckIfRemoteBranch(repoUrl, repoRef)
+	if err != nil {
+		return fmt.Errorf("check if remote branch failed: %w", err)
+	}
+	if isBranch {
+		command := fmt.Sprintf("git clone --branch %s %s --recursive %s", repoRef, repoUrl, repoDir)
+		return cmd.NewExecutor(title, command).Execute()
+	}
 
-	case CheckIfRemoteCommit(repoUrl, repoRef):
-		// Clone repo.
-		cloneCmd := fmt.Sprintf("git clone %s %s --depth 1", repoUrl, repoDir)
-		if err := cmd.NewExecutor(title, cloneCmd).Execute(); err != nil {
-			return err
-		}
+	// ============ Clone specific tag ============
+	isTag, err := CheckIfRemoteTag(repoUrl, repoRef)
+	if err != nil {
+		return fmt.Errorf("check if remote tag failed: %w", err)
+	}
+	if isTag {
+		command := fmt.Sprintf("git clone --branch %s %s --recursive %s", repoRef, repoUrl, repoDir)
+		return cmd.NewExecutor(title, command).Execute()
+	}
 
-		// Checkout commit.
-		checkoutCmd := fmt.Sprintf("git -C %s checkout %s", repoDir, repoRef)
-		if err := cmd.NewExecutor(title, checkoutCmd).Execute(); err != nil {
-			return err
-		}
+	// ============ Clone and checkout commit ============
+	command := fmt.Sprintf("git clone --no-checkout %s %s", repoUrl, repoDir)
+	if err := cmd.NewExecutor(title, command).Execute(); err != nil {
+		return fmt.Errorf("clone git repo failed: %w", err)
+	}
 
-		// Update submodules.
-		submoduleCmd := fmt.Sprintf("git -C %s submodule update --init --recursive", repoDir)
-		if err := cmd.NewExecutor(title, submoduleCmd).Execute(); err != nil {
-			return err
-		}
+	// Checkout repo to commit.
+	command = fmt.Sprintf("git reset --hard %s", repoRef)
+	executor := cmd.NewExecutor(title+" (reset to commit)", command)
+	executor.SetWorkDir(repoDir)
+	if err := executor.Execute(); err != nil {
+		return fmt.Errorf("reset --hard failed: %w", err)
+	}
 
-	default:
-		return fmt.Errorf("ref %s is not a remote branch, tag or commit", repoRef)
+	// Update submodules.
+	if fileio.PathExists(filepath.Join(repoDir, ".gitmodules")) {
+		command = "git submodule update --init --recursive"
+		executor = cmd.NewExecutor(title+" (clone submodule)", command)
+		executor.SetWorkDir(repoDir)
+		if err := executor.Execute(); err != nil {
+			return fmt.Errorf("update submodules failed: %w", err)
+		}
 	}
 
 	return nil
@@ -87,26 +99,38 @@ func UpdateRepo(title, repoRef, repoDir string, force bool) error {
 	var commands []string
 	commands = append(commands, "git reset --hard && git clean -xfd")
 
-	switch {
-	case CheckIfRemoteBranch(repoDir, repoRef):
+	updateRepo := func(commands []string) error {
+		commandLine := strings.Join(commands, " && ")
+		executor := cmd.NewExecutor(title, commandLine)
+		executor.SetWorkDir(repoDir)
+		return executor.Execute()
+	}
+
+	// Update to branch.
+	isBranch, err := CheckIfRemoteBranch(repoDir, repoRef)
+	if err != nil {
+		return err
+	}
+	if isBranch {
 		commands = append(commands, fmt.Sprintf("git fetch origin %s", repoRef))
 		commands = append(commands, fmt.Sprintf("git checkout -B %s origin/%s", repoRef, repoRef))
 		commands = append(commands, fmt.Sprintf("git pull origin %s", repoRef))
+		return updateRepo(commands)
+	}
 
-	case CheckIfRemoteTag(repoDir, repoRef):
+	// update to tag.
+	isTag, err := CheckIfRemoteTag(repoDir, repoRef)
+	if err != nil {
+		return err
+	}
+	if isTag {
 		commands = append(commands, fmt.Sprintf("git tag -d %s || true", repoRef))
 		commands = append(commands, "git fetch --tags origin")
 		commands = append(commands, fmt.Sprintf("git checkout %s", repoRef))
-
-	case CheckIfRemoteCommit(repoDir, repoRef):
-		commands = append(commands, fmt.Sprintf("git reset --hard %s", repoRef))
+		return updateRepo(commands)
 	}
 
-	// Execute clone command.
-	commandLine := strings.Join(commands, " && ")
-	executor := cmd.NewExecutor(title, commandLine)
-	executor.SetWorkDir(repoDir)
-	return executor.Execute()
+	return fmt.Errorf("invalid repoRef: %s", repoRef)
 }
 
 // CherryPick cherry-pick patches.
